training-module/requirements.txt
os==
psycopg2==2.9.9
PyPDF2==3.0.1
logging==0.4.9.6
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
uuid==
pinecone-client==2.2.4
configparser==6.0.0